Added an "ALL" option for the age directive.  Primarily for testing purposes, but useful for multi-user mailboxes to be sorted by To: address.
Added the "To" option for archive range directive.  Will use the user name (i.e. before the @domain.tld) of the To address for the filing location.  If there are multiple addresses in the To: email header, an attempt is made to match locate one of the To: addresses to any email addresses found in the Received: mail headers.  This will probably work for legitimate email, but probably break for heavily falsified mail headers from spammers.  YMMV

Added the "From" option for archive range directive, but will re-work.  Current version files by the root domain name of the From: address, which is only useful in certain circumstances.  The problem is all messages From a @gmail.com address will go into the same folder.  Not much use there.  Questioning how useful a generic From sort option really is.

Added a check for auth=netrc to look for a .netrc entry by imaphost and username, to allow for multiple accounts from the same imap host.

Updated the example archiveimaprc file.